Break out SmileComm class
